[zx][kazoo] Introduce @inout
This CL introduces the `@inout` attribute for syscall parameters, which
signifies that although they might appear in the input 'request' struct
in this representation, they are intended to be mutable 'in-out'
parameters in the syscall interface proper.

The attribute replaces the use of the magic "Mutable"-prefixed aliases
that effectively signified the same thing. Ultimately, @inout is an
expedient placeholder until we are able to model things better (or at
the very least until we have deprecated kazoo).

Bug: 110294
Change-Id: Id4457543016d3cf69f1511ebb40b459fcbdad20a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/726630
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>