[component][fuzzing] Add timeout to Fuzzer::stop
This CL adds a timeout to the method that implements
`fuchsia.fuzzer.Manager/Stop`. If a fuzzer is hung, e.g. stuck in a
blocking read, it may not be able to handle the closure of the
ControllerProvider channel from the registry and exit.

Without this change, the manager will be stuck waiting for the
fuzzer's output to finish, and the client to the manager will be
stuck waiting for a response to the call to `Stop`.

With this change, the manager will notice the fuzzer's failure to
exit and ask the test_manager to kill it, ensuring an eventual response
to the client.

Change-Id: I395a0a33122cd71080a1c13f2201f9afa7201b1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/741523
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>