[fshost] respect fvm_ramdisk flag
When the fvm_ramdisk flag is provided, blob and data are required to be
on a ramdisk. Any fvm partitions not on a ramdisk will still be bound,
and zxcrypt will be bound if appropriate for the data format, but
filesystems on them won't be launched.

Bug: 87511
Fixed: 109334
Test: fx test fshost_integration_tests
Change-Id: Iae15eae660c27377520bc20509dde05196c4d31f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739529
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>