[swd] Generate new devhost roots to remove keyid_hash_algorithms
`keyid_hash_algorithms` is a nonstandard extension to TUF metadata,
which was in place when we generated 7.root.json. rust-tuf warns when we
see metadata signed with unknown key ids, so when we switch over to
`package-tool repository publish` the build will always emit this
warning.

To avoid that, this generates 8.root.json and 9.root.json to allow us to
transition to metadata that no longer references this field and avoid
the warning.

Fixed: 111783
Change-Id: Id3c2bd196c26f4336ac99a009750e06a8839cd9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739776
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>