[debugdata-item] Add kernel-zbi-test for debugdata
Add a test that verifies that arbitrary debugdata items are plumbed
all the way through to userspace and are correct.

The test uses the turducken cuckoo zbi to use a test that appends
two debugdata items, and then boots into zircon for a userspace test
to check that the contests are exactly what is to be expected.
Contents, order and quantity both verified.

Test: fx run-zbi-test early-boot-debug-data-item-zbi-test

Change-Id: I2852eaee66b950029b56f010d9176d4e0441c131
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727793
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>