[fxfs] Prevent truncate from exceeding transaction size
This splits up truncation so that it is performed over multiple
transactions. If we do that, then we need to handle the case where not
all of the transactions get replayed. To address that, we can reuse the
graveyard: if we need to use more than one transaction, we insert a
graveyard entry with the first transaction and then remove it for the
last entry.

Fixed: 110843
Change-Id: I6c031c4b04ad670cea4cd304715dee38150334e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/740082
Reviewed-by: Chris Tam <godtamit@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>