[docgen] Improve docgen links.
Create links in the toplevel reference for the library (previously the
linkifier was not called for this text).

Format cross-references as HTML links rather than markdown links. This
allows us to avoid linking the (...) section which may follow the name.
It looks incorrect to link that in cases where parameter values are
specified.

Adds more usage advice to the README.

Adds () to the functions in the index page. This makes them look more
like function calls. Adds struct/class/union tags to those index items
to clarify what they are.

Change-Id: I06dc212a8f9e9c169ae8815a2cdca85c76c7962a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742227
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>