[amlogic-display] Fix initialization of TV070WSM_FT LCD
Until change Ibf2bda6ceb05f3192125bc03bf5124e760399857 , opcode
0xFD was treated as identical to a write command. With that patch it's
treated as kDsiOpDelay, which is a complicated sleep, so most of the LCD
initialization was being ignored.

To fix this, refresh the init codes for this panel. All the 0xFD entries
are converted to 0x23, which is a short write command.

Bug: b/252224823
Bug: 112367
Tested: "display-tweak panel --power off && display-tweak panel --power on" on Nelson with panel
Tested: testing on an astro with the panel
Change-Id: I3a8a022ec29a821c2cfa1a87975b547e5e62c078
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/741971
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: Victor Costan <costan@google.com>