[packet] Simplify options parsing errors
There are two types of errors that can be encountered while parsing a
sequence of options: Errors in parsing the sequence format (kind field,
len field, number of data bytes) and errors in parsing the data field
itself, which is handled by user-supplied code. Previously, to express
both of these types of errors, we used an enum like this:

  enum OptionParseErr<E> {
      Internal,
      External(E),
  }

This had a few downsides. First, the names are unintuitive. Second, it
results in multi-layer matches. Third, if the external error type can
already express a concept like "malformed sequence of options", it
requires user code to map the `OptionParseErr::Internal` variant to the
equivalent `E` error value.

In this commit, we switch to using only a user-provided error type. This
type implements a trait which provides an associated constant that can
be used by our framework code to generate what previously would have
been considered "internal" errors. This makes the API slightly simpler
(the `Options` parsing code returns `E` errors instead of
`OptionParseErr<E>`) and eliminates the problems mentioned above.

Change-Id: I34bd0d97d3086af35fc58622fbb4cd38cb2a6f32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742006
Reviewed-by: Alex Konradi <akonradi@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>