[packet] Fix options parsing bug
Previously, when parsing an options format with a multi-byte kind field,
if we encountered a partial kind field (e.g., 1 byte where 2 are
expected), we would spuriously treat this as equivalent to having 0
bytes left. This commit fixes that bug and adds a test for it.

Change-Id: I468121f5712b73c4e704460f580f166c876ee7d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/741465
Fuchsia-Auto-Submit: Joshua Liebow-Feeser <joshlf@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>