[docs] Removing link to historic /docs page
This removes the link to historic docs/ page referenced
by a git hash. The doc_checker contraints will disallow links
to old versions of files in the //docs directory. If this needs to
be done in the future, we can revisit the constraint, or with the malice
of forethought move the markdown file out of //docs and link to it in
the new location.

Change-Id: I67851c9d8a478a61bfc31b64e924b7297a64c46f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742864
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Claudio Cherubino <ccherubino@google.com>