[kernel][vm] Remove redundant user_child_count_
The user_child_count_ in VmObject is a remnant from before VmCowPages
was introduced. Before the separation of logic offered by VmCowPages, it
was possible for a hidden VmObjectPaged to exist, which meant the child
count as seen by the user and the internal child count maintained by the
VMO hierarchy could drift apart. This is no longer possible as each
VmObjectPaged has a corresponding dispatcher and so the length of the
children_list_ is the true count of children as seen by userspace.

A CQ run of fxr/742865 verified that user_child_count_ could indeed be
used interchangeably with children_list_len_ by asserting equality of
the two in various places.

Change-Id: Ie39c2a3065958152676ecd2bcc59333815afcf1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742869
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>