[fx][flash-remote]: check if ffx is built locally
fetch_or_build_tool is slow for most workflows, especially if the tool
is already built.

If the remote OS and the host OS are different in fetch_or_build_tool,
then use the fx-cmd-locator functions to get the host-tools dir (if it
exists) and check if the target tool exists there. If so, we can skip
building or fetching the tool and return the path to the host tool.

Otherwise, go through the fetch_or_build_tool workflow as normal.

This also adds the remote OS's operating system in the cache file
(colon delimited). Long term this will be passed to
`fetch_or_build_tool`

BUG: 79023
Change-Id: I450e30064b8ec141775b1695cc553348e20ed95a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/732459
Reviewed-by: Gary Boone <gboone@google.com>
Reviewed-by: David Pursell <dpursell@google.com>
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>