[debugger] Clean up obsolete code.
Removes code protected by the LLVM_USING_OLD_PREBUILT #define. The roll
this helped support is complete so we don't need this any more.

Disables the code to set this define, but leaves it in the tree for use
next time we do a roll.

Renames DwarfDieScanner2 to DwarfDieScanner (I think we used to have an
old version of this).

Change-Id: I0d415e61c89ae4f67bfc701e456ae035e0c1f84d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742873
Reviewed-by: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>