[audio][drivers] Add delays hanging gets
Allow for drivers to report updates to delays reported
for synchronizing presentation across the audio subsystem.
Also we add the ability to report the delay added by
hardware or software that is abstracted by the driver as a
zx.duration rather than the FIFO size, this makes the
information more generic for drivers that don't abstract
hardware with a FIFO, examples are a Bluetooth output
driver and a PDM input driver with software in the data path.

Fixed: 109820

Test: fuchsia-audio-device-output-tests
Test: sa-unittest
Test: aml-g12-tdm-test
Test: aml-g12-tdm-dai-test
Test: usb-audio-test
Change-Id: Ie40b55e9a2dc782bad3198160e171ab633ff7a86
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/727151
Reviewed-by: Martin Puryear <mpuryear@google.com>
API-Review: Marie Janssen <jamuraa@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>