[sdmmc][sdhci] Add support for unowned VMOs
Add support for scatter-gather requests using VMOs that have not been
pre-registered. This is required for eMMC support as the block driver
does not currently register VMOs. Also change PerformTuning to use
unowned VMOs with the new interface so that the old interface can be
removed.

Bug: 106851
Test: Booted Atlas 10 times, eMMC tuning succeeded @ HS400 each time
Test: fx test sdhci-test
Change-Id: Ib2707c7123a5de844e234e665de77c702cd3122f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739605
Reviewed-by: Andres Oportus <andresoportus@google.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>