[build] adds fuchsia_package_with_default_component
This template simplifies creating a packaged component with a defacto
standard component manifest, default.cm. Developers could implement this
convention by explicitly naming their component `default` but this
template codifies the defacto standard so it is more convenient and
correctly implemented.

When hermetic tests are migrated from single-package implementations to
structured subpackaged component implementations, the subpackaged
components can (and perhaps should) use this template to standardize
subpackage URLs, avoiding the redundancy of URLs that use the same name
for the package part and the component part, and ensuring consistency.

Also, as a possible follow-on, if this standard `default.cm` component
manifest name is widely adopted, it should be convenient for Fuchsia to
accept component URLs that end with the URI `path` (which, for
subpackaged component references, would be _only_ the `path`), and infer
the URI fragment is `#meta/default.cm` without requiring it explicitly.

Change-Id: If2246446bf3427c1e532c8787d3edf4406c86cc7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743128
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Rich Kadel <richkadel@google.com>