[kernel] Print OOPS when secondaries fail to boot
Print a KERNEL_OOPS if we fail to boot any of the secondary CPUs.

The motivation is to ensure that CI/CQ detects the failure while
allowing the system to proceed so that logs and other diagnostic data
can make it out of the system.

Fixed: 112387
Run-All-Tests: true
Change-Id: I325c2f5f9ebb3ddc7bd1a90eb1e2f3a7b87a5304
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743503
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>