[gcs] Default to new OAuth2 non-oob flow
The support for OAuth2 oob is being removed. To continue to support
OAuth2 this CL changes from oob to a local loopback http listener
to receive the authorization code (complying to itef rfc8252).

The old behavior of using the oob flow can be accessed with the
--oob_auth switch. This is transitional and may be removed in the
future.

Bug: 108402
Change-Id: I53a8ae744ce01cd6a737eedf4be76424c040e62c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/723048
Commit-Queue: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>