[fs_management] call shutdown in Drop
When we drop a serving filesystem instance, it's expected to attempt a
graceful shutdown of the filesystem. Fix the Drop impls in the rust
fs_management to actually call shutdown for all scenarios. Also, use the
synchronous proxy so that we know it's shut down once we finish
dropping.

Test: fx test fs_management_tests
Change-Id: If4f20e623590c0bda20c782b91d17ac364d69e12
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742935
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>