[netstack] Throttle dhcp logs
Applies the following heuristic to a handful of particularly spammy
log paths:
1. If we have logged, or attempted to log, this line within the past 70
   seconds, swallow the log.
2. However, if we have been swallowing the log for at least an hour,
   actually log the line (along with how many times it's been
   suppressed).
3. Additionally, we manually reset the log throttling state if we
   successfully complete a transaction.

The hope is this will silence uselessly repetitive logs that we didn't
receive a DHCP offer (especially in scenarios where we will never
receive a DHCP offer due to the lack of a DHCP server on the network,
e.g. in cases where we are running in an emulator), while preserving
sufficient debugging signal to understand the state of the DHCP client
from device snapshots.

Bug: 112043
Change-Id: I4bf3df8abebee33d0cc2c0b06770e9e33f97ddbe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/740689
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>