[diagnostics-hierarchy] Simplify Trie
Our Trie implementation was too generic and assuming we'd always be
storing a vector of values at each node.

In the (only) use case of the Trie we store a single value, so the Trie
can be simplified significantly.

This also simplifies the DiagnosticsHierarchy iterator by no longer
relying on the removed types.

Run-All-Tests: true
Bug: 92015

Change-Id: Id20c2ca5f2c2730f94c13642da434938b86c7aac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743422
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>