[fshost][assembly] Move compilation of fshost component to assembled_system()
To allow assembled_system() to be able to properly select the features
configured in fshost, assembled_system() and fshost_cml() need to use
the same options.  This is easiest of it's just moved to within
assembled system itself, which prepares it for use with
product assembly.

Change-Id: I58c4f4439ae89b265f6c1640518082b3391866f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742562
Reviewed-by: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>