[check-licenses] gn package improvements.
The output of "gn gen" contains a list of everything that is defined in
the build. The list of things that are "defined" is vastly larger than
the list of things that are actually used / depended on in the build.

This CL changes check-licenses to now only look at the dependencies of
the //:default target, which (after talking with some folks) appears to
be the target I am most interested in for licensing. This CL adds code
to analyze that target, analyze all dependencies of that target
(recursive), and analyze any inputs that those targets may define.

The result is a significantly smaller NOTICE file, but I believe this
to be more accurate. Using `fx gn path` on a handful of now missing
targets, I confirmed that they were not a part of the build graph.

Change-Id: I693394f158834c9a6067ce8157ff6faa77a6b883
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743247
Commit-Queue: Jerry Belton <jcecil@google.com>
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>