[debugger] Add a simple bytecode machine.
This is a simplest-possible bytecode stack machine. It uses the existing
ExprValue type for the stack type. Most of the machine logic is actually
implemented in callbacks that are registered as machine operations that
take a certain number of arguments. This leaves the machine to implement
only the higher-level control flow.

Using a bytecode machine for the control flow in expressions will
eliminate the deeply nested callbacks with the code written in reverse
order, with every step being potentially asynchronous. This machine
allows synchronous operations to stay synchronous without generating
super deep stacks.

This is not hooked up to anything yet.

Change-Id: I87c9be4f447177db93fcb3a1a2da09a8f5c0d231
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739606
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>