[pkg-cache] RetainedIndex supports subpackages
Before, either a package's meta.far was cached and so all its
required blobs were known, or the meta.far was not cached and
none of the required blobs (besides the meta.far itself) were
known.

With subpackages, the meta.far can be cached but not all required
blobs known if the meta.far of a subpackage is not cached.

Bug: 104909
Change-Id: I8c5f5af1d8fdc133e6af2b5e6f4741fc8bce7987
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742368
Reviewed-by: John Wittrock <wittrock@google.com>
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>