[netstack3] Add method to set RingBuffer capacity
Add the ability to set the target capacity on a RingBuffer without
losing any of the already-written or unread bytes in the buffer. If the
new capacity is an increase, the change will be applied immediately.
Otherwise the target capacity will be recorded, to be performed later
once read operations have cleared enough space in the buffer.

Bug: 110625
Change-Id: I4d42b237bc0ef60286fa63a36ce1ad63e83a72ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742644
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>