[net-types] Make IP type a param for GenericOverIp
Hoist the target IP type for GenericOverIp up so that it is part of the
GenericOverIp trait implementation. The result is that the
implementations of GenericOverIp can have trait bounds for the IP type
that are not just `I: Ip`. This allows GenericOverIp and Ip::map_ip to
be used with types from IP extension traits.

Bug: 111992
Change-Id: Idc06675a94d6a30393e25ca1f57f6c5aa6f7e0ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743486
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>