[starnix] Update restricted mode debug addr
This change uses the debug address of Starnix itself for each process it
creates. This allows the debugger to get the correct module lists for
Starnix itself when using restricted mode.

Note that this is not always desirable. Work is being done to ensure
that the debugger can support module lists for both the restricted and
the normal mode code.

Bug: 101486

Change-Id: I040c452620dadb7460b4f203ef783f473cfdabec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743343
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>