[devicetree] Add MatcherScanResult
In order to have an upperbound in the number of scans we might need
to perform over a devicetree, we introduce MatcherScanResult<N>.

When a matcher may require more than one scan, it express so by
providing an overload that returns MatcherScanResult<N> where N
is the maximum number of scans needed before a Matcher reaches
completion.

Following this, for a set of n matchers, the maximum number of
scans is determined as max{ N_0, ...., N_n-1} where N_i is
MatcherScanResult<N> for matcher i or if the matcher provides
the MatcherResult overload N_i is 1.

Bug: 78675
Test: fx test -v -o devicetree-tests
Change-Id: I1760300895e51480a64cbf7691a4bb3d18c04b5b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739656
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>