[devicetree] Requested Scans enforcement
Previously a matchers state was only checked if they had completed
before the maximum number of scans requested were completed for an early
exit. (E.g. a matcher requests 100 scans, but is done after the first
one).

This allows matchers to provide incorrect information when combined with
the requested scans by other matchers. (E.g. matcher 1 requests 1 scan,
and matcher 2 requests 2 scans, if matcher 1 is not done after 1 scan,
but it is done after 2, it is considered success).

This change, enforce that if a matcher requests N scans, it must be done
by the end of the given scan, otherwise return the index of the matcher
in the matcher set, so a user can debug the respective matcher.

Test: fx test -v -o devicetree-tests
Change-Id: Ia264d8165d68ec71fa1a0b26aa948de2d67d16bb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/724223
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>