[audio][drivers][tas58xx] Allow no fault GPIO
Some product configurations may have the FAULT pin either
not connected to a GPIO input pin or used as an ADR (address)
pin and hence not available to the driver to be used for fault
detection. Allow these configurations by not failing with an
error if no "gpio-fault" GPIO is provided to the driver.

Bug: 107645

Test: tas58xx-test
Test: audio-driver-ctl tone in Pinecrest with subsequent CLs.

Change-Id: Iadff31759b6bd5ff099922ca2c814ee8c3afc321
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744042
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>