[cm] Fix broken "scope"
Previously, we used a broken heuristic to duplicate
the logic that librouting used to permission-check
"scope" in CML manifests. This resulted in the incorrect
offer declarations being used to filter events.
This CL fixes this by no longer duplicating this logic
and relying on librouting to give us this information directly.

Run-all-tests: true

Fixed: 112416

Change-Id: Ic99bdef46d4ac87f02fa3f331d7bedeff585087f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743766
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Gary Bressler <geb@google.com>