[overnet] Fix fallback logic from CSO to legacy
In legacy overnet, we can create a Peer for a node we have no route to.
The peer will fail five seconds later unless a route appears. CSO peers,
by contrast, can only be accessed for nodes we already know we can
connect to.

Our logic for supporting both protocols was "try to get a CSO peer, and
if that fails get a legacy peer." This was broken as, if the network
hadn't settled, and we hadn't seen all of our CSO peers yet, we'd fail
to get a CSO peer immediately, and then return a legacy peer, which
would sit around for a bit and give the legacy network a chance to
quiesce.

This patch lets us keep trying to look for CSO peers while the legacy
network is waiting for a route. As a side effects, a few calls that used
to be relatively quick awaits might now block for a bit longer, because
they have to wait for peers to validate rather than return them
immediately. This shouldn't be too much of a problem, and it gets the
integration tests passing.

Change-Id: I64fec0c5a9bd8dbdfdb574e51746bdb0abe87c44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744804
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>