[make-fuchsia-vol] Prefix images with fuchsia build dir
`make-fuchsia-vol --fuchsia-build-dir ...` will look up the images
out of the build directory's `images.json` file and use them if
the paths were not specified on the command line. However, the
image paths in `images.json` are relative to the build directory.
This meant that this option could only be specified in the build
directory.

This patch fixes the problem by prefixing the `images.json` paths
with the build directory.

Change-Id: I2bffee86a5cb6a371a29da83ee1ee79f4685ad45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/744882
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>