[debugger] Use inline calls for line matches.
This works around a clang bug
https://github.com/llvm/llvm-project/issues/58483
which is that it doesn't emit line table information for the source of
an inline function call if the line consists only of that inline call.
This is fairly easy to trigger (just put an inline call on a line by
itself) and means that you can't set a breakpoint on that line (at least
in optimized code).

This works around the problem by taking all functions that match the
normal file/line lookup, and adding in the inline call information as
potential line matches.

This produces slightly surprising result as the breakpoint set is shown
to match the first line of the inline call (since this is where the
actual address matched to), but everything works as expected.

Bug: 112203

Change-Id: I45438fce4bb5622d41547cea75af544d6cd74e26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743902
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>