[analytics][cpp] Migrate storage to $XDG_DATA_HOME
Ananlytics status/states were stored in $HOME/.fuchsia/metrics,
which is now considered a bad practice. We are moving them into
$XDG_DATA_HOME/Fuchsia/metrics.

If the new location does not exist and the old location is found,
the migration will move the old folder to the new location and create a
symlink at the old location. Otherwise, no change is needed.

Bug: 100850
Change-Id: Id83f95ee73a1e3c229b7cbfae30a9a161cbb9f03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743245
Commit-Queue: Rundong Du <rundongdu@google.com>
Reviewed-by: Bob Evans <bobevans@google.com>