[kernel][arm64][mmu] After modifying PT entries, issue a DSB instead of DMB
DSB ISH (and the slightly weaker ISHST) is necessary to ensure the
previous writes to any page table entries are flushed out of the core
and visible by other processors and page table walkers.

Previously was using DMB which is actually insufficient since it only
guarantees the order the entries are flushed into the memory subsystem,
not that they actually are.

In most cases it's sufficient to use a DSB ISHST since it ensures only
previous writes were barriered, and reads by the cpu can go around it.
This is okay because the code wont access any of the new mappings
directly until at least exiting the mmu code.

Change-Id: Ib8c3111f570896909b933ca1a09ec02b996d4914
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/669549
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>