[debugger] Emit bytecode for the bytecode machine.
Emits bytecode from the ExprNodes to run in the new bytecode stack
machine.

No attempt was made to share logic between the AST execution model and
the bytecode model. This resulted in significant code duplication
between the two models. The reason is that if successful, the AST mode
code will be deleted, and if unsuccessful, the bytecode mode code will
be deleted. Since this transition should not take very long, duplicating
the code allows one or the other mode to be deleted with minimal
disruption.

This code is not yet hooked up except for tests. The ExprNode tests now
run each variant of the test in the old "Eval" method that runs on the
AST, as well as the bytecode version.

Change-Id: I08311e64edcd2706b92f0b70563ca8d0e8be961f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/739607
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>