[stream] Fix crash when iovec is resized
This change fixes a crash that occurs when a `zx_iovec_t` passed into
a stream operation is resized while the operation is ongoing.

This crash only occurred on debug builds due to a DEBUG_ASSERT on the
status was being hit.

Fixed: 111666

Run-All-Tests: true
Multiply: core-stream:25
Test: core-stream

Change-Id: I124d9600bc6a27046877464d4cc24ac37fc1d47d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/742931
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Chris Tam <godtamit@google.com>