[virtualization] Parameterize net device tests
This runs the tests against the C++ and Rust components. This skips the
RX test for now, as the device doesn't yet support that.

This also does a cleaner shutdown (which the C++ device doesn't seem to
do). This stops some asserts firing when the component terminates, which
only happens for the Rust device because we support stopping the
component on error.

Bug: 95485
Change-Id: I0f9889eda21e60a340fff0efb142e08c66015eca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/743964
Commit-Queue: David Hastings <dahastin@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>